# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Welcome to the HavenStay Backend API documentation! 🚀\n\nHavenStay is a modern, high-performance Domain-Driven Design (DDD) backend built in Go. It empowers Hosts to manage luxury properties and Guests to seamlessly book remote getaways.\n\n### Key Features:\n- **🏘️ Property Management**: Full CRUD lifecycle for short-term rentals and luxury homes.\n- **👩‍💼 Host \u0026 Guest Profiles**: Independent onboarding and profile management domains.\n- **📅 Booking Engine**: Real-time reservation creation.\n- **❤️ Wishlists**: Guests can curate lists of favorite properties.\n\n**Built With**: Go 1.21+, Gin (REST), gRPC, Protobufs, MySQL 8, GORM.",
        "title": "HavenStay - Hotel \u0026 Room Booking API",
        "termsOfService": "https://havenstay.localhost/terms",
        "contact": {
            "name": "HavenStay Platform Support",
            "url": "https://github.com/AshutoshKY/havenstay-backend",
            "email": "api-support@havenstay.local"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/lists": {
            "get": {
                "description": "Fetches the complete array of bookmarked properties saved under a specific Guest account.\nActs as the primary data source for the \"Saved Homes\" UI tab.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wishlists"
                ],
                "summary": "View Guest Wishlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The target Guest's unique Identifier.",
                        "name": "guestId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array wrapping the nested Property details saved by the user.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.GetWishlistResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Bookmarks a specific property to a Guest's curated wishlist for later viewing or booking.\nUseful for encouraging users to return to properties they found engaging.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wishlists"
                ],
                "summary": "Save Property to Wishlist",
                "parameters": [
                    {
                        "description": "Payload connecting a Guest ID to a Property ID.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.AddToWishlistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "A confirmation object encapsulating the newly bookmarked item constraint.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.WishlistResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Duplicate bookmark or invalid identifiers.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_user_airbnb-test_api_proto_v1.AddToWishlistRequest": {
            "type": "object",
            "properties": {
                "guest_id": {
                    "type": "integer"
                },
                "property_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_user_airbnb-test_api_proto_v1.GetWishlistResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.WishlistItem"
                    }
                }
            }
        },
...
        "github_com_user_airbnb-test_api_proto_v1.WishlistItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "guest_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "property_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_user_airbnb-test_api_proto_v1.WishlistResponse": {
            "type": "object",
            "properties": {
                "item": {
                    "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.WishlistItem"
                }
            }
        },
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. This field is\nthe nanosecond portion of the duration, not an alternative to seconds.\nNegative second values with fractions must still have non-negative nanos\nvalues that count forward in time. Must be between 0 and 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z. Must\nbe between -315576000000 and 315576000000 inclusive (which corresponds to\n0001-01-01T00:00:00Z to 9999-12-31T23:59:59Z).",
                    "type": "integer"
                }
            }
        }
    }
}